+++ datadog/data_source_datadog_ip_ranges.go
@@ -2,7 +2,7 @@ package datadog
 
 import (
 	"github.com/hashicorp/terraform/helper/schema"
-	datadog "github.com/zorkian/go-datadog-api"
+	datadog "github.com/s33dunda/go-datadog-api"
 )
 
 func dataSourceDatadogIpRanges() *schema.Resource {