+++ datadog/resource_datadog_dashboard_list.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	"github.com/hashicorp/terraform/helper/schema"
-	datadog "github.com/zorkian/go-datadog-api"
+	datadog "github.com/s33dunda/go-datadog-api"
 )
 
 func resourceDatadogDashboardList() *schema.Resource {