+++ datadog/resource_datadog_integration_pagerduty.go
@@ -6,7 +6,7 @@ import (
 	"sync"
 
 	"github.com/hashicorp/terraform/helper/schema"
-	"github.com/zorkian/go-datadog-api"
+	"github.com/s33dunda/go-datadog-api"
 )
 
 // creating/modifying/deleting PD integration and its service objects in parallel on one account