+++ datadog/resource_datadog_integration_pagerduty_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
-	datadog "github.com/zorkian/go-datadog-api"
+	datadog "github.com/s33dunda/go-datadog-api"
 )
 
 // We're not testing for schedules because Datadog actively verifies it with Pagerduty