+++ datadog/resource_datadog_logs_custom_pipeline.go
@@ -3,7 +3,7 @@ package datadog
 import (
 	"fmt"
 	"github.com/hashicorp/terraform/helper/schema"
-	"github.com/zorkian/go-datadog-api"
+	"github.com/s33dunda/go-datadog-api"
 	"strings"
 )