+++ datadog/resource_datadog_monitor.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/hashicorp/terraform/helper/schema"
-	"github.com/zorkian/go-datadog-api"
+	"github.com/s33dunda/go-datadog-api"
 )
 
 const logAlertMonitorType = "log alert"