+++ datadog/resource_datadog_synthetics_test_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
-	datadog "github.com/zorkian/go-datadog-api"
+	datadog "github.com/s33dunda/go-datadog-api"
 )
 
 func TestAccDatadogSyntheticsAPITest_importBasic(t *testing.T) {